# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError:
    properties:
      error:
        type: string
    type: object
  internal_chat.ChangeMemberRoleRequest:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/internal_chat.RoomRole'
        enum:
        - member
        - moderator
    required:
    - role
    type: object
  internal_chat.CreateRoomRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/internal_chat.RoomVisibility'
        enum:
        - public
        - private
    required:
    - name
    type: object
  internal_chat.Invitation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      invited_by:
        type: integer
      invitee_id:
        type: integer
      responded_at:
        type: string
      room_id:
        type: integer
      room_name:
        type: string
      status:
        $ref: '#/definitions/internal_chat.InvitationStatus'
      updated_at:
        type: string
    type: object
  internal_chat.InvitationStatus:
    enum:
    - pending
    - accepted
    - declined
    type: string
    x-enum-varnames:
    - InvitationPending
    - InvitationAccepted
    - InvitationDeclined
  internal_chat.InvitationsResponse:
    properties:
      invitations:
        items:
          $ref: '#/definitions/internal_chat.Invitation'
        type: array
    type: object
  internal_chat.InviteUserRequest:
    properties:
      user_id:
        minimum: 1
        type: integer
    required:
    - user_id
    type: object
...
  internal_chat.MembersResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/internal_chat.RoomMember'
        type: array
    type: object
...
  internal_chat.Room:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/internal_chat.RoomRole'
      visibility:
        $ref: '#/definitions/internal_chat.RoomVisibility'
    type: object
  internal_chat.RoomMember:
    properties:
      joined_at:
        type: string
      role:
        $ref: '#/definitions/internal_chat.RoomRole'
      room_id:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  internal_chat.RoomRole:
    enum:
    - owner
    - moderator
    - member
    type: string
    x-enum-varnames:
    - RoomRoleOwner
    - RoomRoleModerator
    - RoomRoleMember
  internal_chat.RoomVisibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - RoomVisibilityPublic
    - RoomVisibilityPrivate
  internal_chat.RoomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/internal_chat.Room'
        type: array
    type: object
...
  internal_chat.TransferOwnershipRequest:
    properties:
      user_id:
        minimum: 1
        type: integer
    required:
    - user_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Gogo DL real-time chat backend.
  title: Gogo DL API
  version: "1.0"
paths:
...
  /api/v1/invitations/{id}/accept:
    post:
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.Invitation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Accept an invitation
      tags:
      - invitations
  /api/v1/invitations/{id}/decline:
    post:
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.Invitation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Decline an invitation
      tags:
      - invitations
  /api/v1/rooms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.RoomsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: List rooms visible to the current user
      tags:
      - rooms
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_chat.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a room
      tags:
      - rooms
...
  /api/v1/rooms/{id}/invitations:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Invitation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.InviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.Invitation'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_chat.Invitation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Invite a user to a room
      tags:
      - invitations
...
  /api/v1/rooms/{id}/members:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.MembersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: List room members
      tags:
      - memberships
  /api/v1/rooms/{id}/members/{user_id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "204":
          description: Member removed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Remove a room member
      tags:
      - memberships
    patch:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Role payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.ChangeMemberRoleRequest'
      responses:
        "204":
          description: Role changed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Change a member role
      tags:
      - memberships
  /api/v1/rooms/{id}/membership:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Membership removed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Leave a room
      tags:
      - memberships
...
  /api/v1/rooms/{id}/ownership:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ownership payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_chat.TransferOwnershipRequest'
      responses:
        "204":
          description: Ownership transferred
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: Transfer room ownership
      tags:
      - memberships
...
  /api/v1/users/me/invitations:
    get:
      parameters:
      - default: pending
        description: Invitation status filter
        enum:
        - pending
        - accepted
        - declined
        - all
        in: query
        name: status
        type: string
      - default: 50
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_chat.InvitationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_xuanphongtran_gogo-dl_pkg_apperror.AppError'
      security:
      - BearerAuth: []
      summary: List my invitations
      tags:
      - invitations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the access token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"